frequency changed to in

+++ src/main/java/com/github/meshotron2/room_creator/Room.java
@@ -11,13 +11,13 @@ public class Room {
     private final int y;
     private final int z;
 
-    private final long f;
+    private final int f;
 
     private DataInputStream reader;
     private DataOutputStream writer;
     private RandomAccessFile randWriter;
 
-    public Room(String file, int x, int y, int z, long f) {
+    public Room(String file, int x, int y, int z, int f) {
         this.file = file;
         this.x = x;
         this.y = y;
@@ -93,7 +93,7 @@ public class Room {
         final int y = Integer.reverseBytes(reader.readInt());
         final int z = Integer.reverseBytes(reader.readInt());
 
-        final long f = Long.reverseBytes(reader.readLong());
+        final int f = Integer.reverseBytes(reader.readInt());
 
         reader.close();
         return new Room(file, x, y, z, f);
